move some class to new package
