+++ app/src/main/java/com/dicoding/picodiploma/sejiwaproject/LeagueAdapter.kt
@@ -1,5 +1,6 @@
 package com.dicoding.picodiploma.sejiwaproject
 
+import android.content.Intent
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
@@ -7,6 +8,8 @@ import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
+import com.dicoding.picodiploma.sejiwaproject.detail.DetailLeagueActivity
+import com.dicoding.picodiploma.sejiwaproject.model.League
 import com.dicoding.picodiploma.sejiwaproject.view.ItemLeagueUI
 import org.jetbrains.anko.AnkoContext
 
@@ -31,11 +34,15 @@ class LeagueAdapter(private val list: ArrayList<League>) :
         holder.txtName.text = leagues.name
         holder.txtDesc.text = leagues.description
         holder.txtLocation.text = leagues.location
-
+        holder.itemView.setOnClickListener {
+            val detailLeagueActivity = Intent(it.context, DetailLeagueActivity::class.java)
+            detailLeagueActivity.putExtra(DetailLeagueActivity.EXTRA_KEY, list[holder.adapterPosition])
+            it.context.startActivity(detailLeagueActivity)
+        }
 
     }
 
-    class ListViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
+    inner class ListViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
         val imgLogo: ImageView = itemView.findViewById(R.id.league_logo) as ImageView
         val txtName: TextView = itemView.findViewById(R.id.name_league) as TextView
         val txtDesc: TextView = itemView.findViewById(R.id.league_desc) as TextView
